tests/requirements.txt
pytest==7.4.0
pytest-cov==4.1.0
factory-boy==3.3.0
Faker==19.6.2
freezegun==1.4.0
requests-mock==1.11.0
